+++ packages/markdown/stories/Markdown.stories.tsx
@@ -1,4 +1,7 @@
 import type { StoryObj } from '@storybook/react';
+import type { DOMNode, Element, Text } from 'html-react-parser';
+
+import parse from 'html-react-parser';
 
 import { SAMPLE_MARKDOWN } from '../__tests__/constants';
 import { Markdown } from '../src';
@@ -52,5 +55,87 @@ export const Example: Story = {
   args: {
     baseUri: 'https://app.launchdarkly.com',
     source: SAMPLE_MARKDOWN,
+  },
+};
+
+const Link = (props: { href: string }) => {
+  return (
+    <span style={{ backgroundColor: 'red' }}>
+      <a href={props.href}>Hey there!</a>
+    </span>
+  );
+};
+
+type SmartLink = {
+  domain: string;
+  renderer: (props: { href: string; text?: string }) => JSX.Element;
+};
+
+function isAnchorDOMNode(node: DOMNode): node is Element {
+  const element = node as Element;
+  return element?.name === 'a';
+}
+
+function isDOMText(node: DOMNode): node is Text {
+  return (node as Text)?.type === 'text';
+}
+
+function parseLinkFromDOMNode(node: DOMNode) {
+  if (!isAnchorDOMNode(node)) {
+    return;
+  }
+  if (!('href' in node.attribs)) {
+    return;
+  }
+  const { href } = node.attribs;
+  const textNode = node.firstChild;
+  if (!textNode || !isDOMText(textNode)) {
+    return;
+  }
+  try {
+    const url = new URL(href);
+    return { url, text: textNode.data !== href ? textNode.data : undefined };
+  } catch {
+    return;
+  }
+}
+
+export const WithChildrenRender: Story = {
+  render: () => {
+    const parseMarkdown = (renderedMarkdown: string) =>
+      parse(renderedMarkdown, {
+        // The replace option will replace all DOM nodes when the replace function returns a valid React element.
+        replace: (domNode) => {
+          const parsedLink = parseLinkFromDOMNode(domNode);
+          if (!parsedLink) {
+            return;
+          }
+          const smartLinks: SmartLink[] = [
+            {
+              domain: 'launchdarkly.com',
+              renderer: ({ href }) => <Link href={href} />,
+            },
+          ];
+          const smartLink = smartLinks?.find((s) => {
+            return parsedLink.url.hostname.endsWith(s.domain);
+          });
+          if (!smartLink) {
+            return;
+          }
+          return smartLink.renderer({
+            href: parsedLink.url.toString(),
+            text: parsedLink.text,
+          });
+        },
+      });
+
+    return (
+      <Markdown
+        source={SAMPLE_MARKDOWN}
+        allowedTags={['a', '#text', 'p', 'li', 'ol', 'ul', 'b', 'strong', 'i', 'em', 'br', 'code']}
+      >
+        {parseMarkdown}
+      </Markdown>
+    );
   },
 };