+++ packages/select/src/index.ts
@@ -1,0 +1,5 @@
+export type { SelectProps, SelectTriggerProps, SelectTriggerChildrenState } from './Select';
+
+export { Select } from './Select';
+export { SingleSelectTrigger } from './SingleSelectTrigger';
+export { MultiSelectTrigger } from './MultiSelectTrigger';