+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @launchpad-ui/toast
 
+## 0.2.27
+
+### Patch Changes
+
+- Updated dependencies [[`990af4a5`](https://github.com/launchdarkly/launchpad-ui/commit/990af4a5c7490a322604272cf9edc16ce9427cf1)]:
+  - @launchpad-ui/icons@0.6.19
+
+## 0.2.26
+
+### Patch Changes
+
+- [#666](https://github.com/launchdarkly/launchpad-ui/pull/666) [`20439f2d`](https://github.com/launchdarkly/launchpad-ui/commit/20439f2d0533bf574b613d16d5d31eb688cb629f) Thanks [@chasedarkly](https://github.com/chasedarkly)! - Bump z-index values
+
+  [Drawer] Increase z-index from 300 to 400
+  [Modal] Increase z-index from 400 to 500
+  [Popover] Increase z-index from 500 to 600
+  [Snackbar] Increase z-index from 700 to 800
+  [Toast] Increase z-index from 700 to 800
+  [Tooltip] Increase z-index from 600 to 700
+
+- Updated dependencies [[`808902f7`](https://github.com/launchdarkly/launchpad-ui/commit/808902f71c0b410adf86b0ee1683a2d36c4cdb22)]:
+  - @launchpad-ui/tokens@0.4.9
+  - @launchpad-ui/icons@0.6.18
+
+## 0.2.25
+
+### Patch Changes
+
+- Updated dependencies [[`f873c6f3`](https://github.com/launchdarkly/launchpad-ui/commit/f873c6f3af18b55e72883464e9698ea172021c8e)]:
+  - @launchpad-ui/icons@0.6.17
+
+## 0.2.24
+
+### Patch Changes
+
+- Updated dependencies [[`cd1b58c5`](https://github.com/launchdarkly/launchpad-ui/commit/cd1b58c5c117d37b2939fe879606011c49b18ced)]:
+  - @launchpad-ui/icons@0.6.16
+
 ## 0.2.23
 
 ### Patch Changes